Add `suite create` alias
