+++ openforbc_benchmark/cli/suite.py
@@ -6,7 +6,8 @@ if TYPE_CHECKING:
 from sys import stdout
 from typer import Context, echo, Exit, Typer, Option  # noqa: TC002
 
-from openforbc_benchmark.benchmark import BenchmarkSuite
+from openforbc_benchmark.benchmark import Benchmark, BenchmarkSuite, get_benchmarks
+from openforbc_benchmark.json import BenchmarkRunDefinition, BenchmarkSuiteDefinition
 from openforbc_benchmark.cli.benchmark import CliBenchmarkRun
 from openforbc_benchmark.cli.state import state
 
@@ -67,6 +68,89 @@ def find_suite(search: str, search_path: str) -> "Optional[BenchmarkSuite]":
     )
 
 
+def prompt_benchmark_run(benchmarks: "List[Benchmark]") -> BenchmarkRunDefinition:
+    from inquirer import checkbox, list_input
+
+    benchmark_name = list_input(
+        "Select the benchmark", choices=[benchmark.get_id() for benchmark in benchmarks]
+    )
+    benchmark = next(
+        benchmark for benchmark in benchmarks if benchmark.get_id() == benchmark_name
+    )
+
+    presets = benchmark.get_presets()
+    selected_preset_names = checkbox(
+        "Select some presets (select with <spacebar>)",
+        choices=[preset.name for preset in presets],
+        default=benchmark.get_default_preset().name,
+    )
+
+    if not selected_preset_names:
+        echo("ERROR: No preset selected", err=True)
+        raise Exit(1)
+
+    selected_presets = [
+        preset for preset in presets if preset.name in selected_preset_names
+    ]
+
+    if not selected_presets:
+        echo(
+            "ERROR: Couldn't find any of presets \""
+            f'{", ".join(selected_preset_names)}"',
+            err=True,
+        )
+        raise Exit(1)
+
+    return BenchmarkRunDefinition(
+        benchmark.get_id(), [preset.name for preset in presets]
+    )
+
+
+def suite_definition_tool() -> None:
+    from inquirer import confirm, text
+    from json import dumps
+    from os.path import exists, join
+
+    benchmarks = list(get_benchmarks(state["search_path"]))
+
+    suite_name = text("Suite name")
+    suite_description = text("Suite description")
+
+    benchmark_runs: "List[BenchmarkRunDefinition]" = []
+
+    while True:
+        benchmark_runs.append(prompt_benchmark_run(benchmarks))
+        if not confirm("Do you want to add another benchmark run?"):
+            break
+
+    suite = BenchmarkSuiteDefinition(suite_name, suite_description, benchmark_runs)
+
+    json = dumps(suite, indent=2, default=suite.serialize)
+    echo("Previewing JSON...")
+    echo(json)
+
+    if confirm("Save definition to suites folder?", default=True):
+        while True:
+            filename = text("filename (.json will be added if missing)")
+
+            if not filename:
+                echo("Filename can't be empty!")
+                continue
+
+            if not filename.endswith(".json"):
+                filename += ".json"
+
+            path = join("suites", filename)
+
+            if exists(path) and not confirm("File already exists, overwrite?"):
+                continue
+
+            with open(path, "w") as file:
+                file.write(json)
+
+            break
+
+
 app = Typer(help="List, inspect and run benchmark suites")
 
 
@@ -114,6 +198,12 @@ def get_suite_info(suite_name: str) -> None:
             tablefmt="simple",
         )
     )
+
+
+@app.command("create")
+def create_suite() -> None:
+    """Create a suite interactively."""
+    return suite_definition_tool()
 
 
 @app.command("run")
